+++ Products/membrane/plugins/usermanager.py
@@ -155,6 +155,14 @@ class MembraneUserManager(BasePlugin, Cacheable):
             # Title is a ZCTextIndex, we don't need to look for exact_match.
             query['Title'] = keywords['fullname']
 
+        if not query:
+            # The query is empty.  If we continue, that would lead to
+            # returning all members, which is not what we want here.
+            # Note that if in Plone you click 'Show all users' in the
+            # users panel, that leads to three queries.  At least one
+            # of those will not be empty, so all users will be listed
+            # anyway, which is good.
+            return []
         if sort_by is not None:
             if sort_by == 'login':
                 query['sort_on'] = 'getUserName'